Finalize and clean up.
Added comments, moved the ajax call backs up to the events. Created a
list to store the urls after proccessing, this way you can list all the
url for easier consumption. Removed the geturl template, since I moved
the form to the layout template. I wanted to be able to continously
process urls on the same page.